refactor: move contstant values to config
