+++ config.py
@@ -1,0 +1,47 @@
+import os
+from selenium import webdriver
+from selenium.webdriver.common.desired_capabilities import DesiredCapabilities
+
+class Config:
+    SLEEP_INTERVAL = 15
+
+    CATEGORIES_DICT = {
+        'Parque': ['parque', 'parques de diversões', 'ar livre', 'bonde', 'natureza', 'trilha'],
+        'Museu': ['museus', 'locais históricos', 'arte', 'religioso', 'igreja'],
+        'Cinema': ['cinema'],
+        'Shopping': ['shopping', 'lojas especializadas', 'outlet', 'calçadões'],
+        'Bar': ['bar', 'bares', 'cervejarias', 'balada'],
+        'Restaurante': ['restaurante', 'feira'],
+        'Show': ['show'],
+        'Biblioteca': ['biblioteca'],
+        'Estádio': ['estádio', 'arenas', 'esportivo'],
+        'Jogos': ['jogos', 'boliche', 'entretenimento', 'stand de tiro'],
+        'Teatro': ['teatro']
+    }
+
+    CITIES_LIST = {
+        'Santo André': 'https://www.tripadvisor.com.br/Attractions-g303624-Activities-Santo_Andre_State_of_Sao_Paulo.html',
+        'São Bernardo do Campo': 'https://www.tripadvisor.com.br/Attractions-g303626-Activities-a_allAttractions.true-Sao_Bernardo_Do_Campo_State_of_Sao_Paulo.html',
+        'São Caetano do Sul': 'https://www.tripadvisor.com.br/Attractions-g1162161-Activities-Sao_Caetano_do_Sul_State_of_Sao_Paulo.html',
+        'Diadema': 'https://www.tripadvisor.com.br/Attractions-g780021-Activities-Diadema_State_of_Sao_Paulo.html',
+        'Mauá': 'https://www.tripadvisor.com.br/Attractions-g2342768-Activities-Maua_State_of_Sao_Paulo.html',
+        'Ribeirão Pires': 'https://www.tripadvisor.com.br/Attractions-g2343028-Activities-Ribeirao_Pires_State_of_Sao_Paulo.html',
+        'Rio Grande da Serra': 'https://www.tripadvisor.com.br/Attractions-g2346575-Activities-Rio_Grande_Da_Serra_State_of_Sao_Paulo.html'
+    }
+
+class SeleniumConfig():
+    def __init__(self):
+        chrome_options = webdriver.ChromeOptions()
+        chrome_options.add_argument('--disable-gpu')
+        chrome_options.add_argument('--no-sandbox')
+        chrome_options.add_argument('--headless')
+
+        if os.getenv('CRAWLER_ENVIRONMENT') == 'dev':
+            self.driver = webdriver.Chrome(os.getenv("CHROME_DRIVER_PATH"), chrome_options=chrome_options)
+        elif os.getenv('CRAWLER_ENVIRONMENT') == 'sandbox':
+            self.driver = webdriver.Remote(command_executor='http://localhost:4444/wd/hub', chrome_options=chrome_options, desired_capabilities=DesiredCapabilities.CHROME)
+        elif os.getenv('CRAWLER_ENVIRONMENT') == 'production':
+            chrome_options.add_argument('--disable-dev-shm-usage')
+            chrome_options.binary_location = os.getenv("GOOGLE_CHROME_BIN")
+
+            self.driver = webdriver.Chrome(executable_path=os.getenv("CHROME_DRIVER_PATH"), chrome_options=chrome_options)
